Fix atomic_t alignment on aarch64: use word-sized value field
On aarch64, atomic_t.value was unsigned int (4 bytes) which caused
field misalignment in packed structures (batch_t, tqnode_t, sched_t).
With __attribute__((packed)), the 4-byte atomic field followed by
8-byte pointer fields resulted in pointers at non-8-byte-aligned
offsets (e.g., tqnode_t.scheduler at offset 36 instead of 40).

This misalignment caused the scheduler timer queue traversal to read
corrupt pointer values, producing NULL pointer dereferences (SIGSEGV
at address 0xC) when processing sprite-heavy occade demos.

Fix: use unsigned long (8 bytes on aarch64) for atomic_t.value.
This maintains alignment of subsequent fields in packed structures.

The remaining occade failures are now SDL/Cocoa threading errors
(nextEventMatchingMask must be called from Main Thread) which is
a macOS-specific SDL requirement, not a 64-bit port issue.

+++ runtime/ccsp/include/aarch64/atomics.h
@@ -33,10 +33,14 @@
 
 typedef struct _atomic_t atomic_t;
 struct _atomic_t {
+	/* Use 64-bit value on aarch64 to maintain alignment of
+	 * subsequent fields in packed structures (batch_t, tqnode_t).
+	 * With unsigned int (4 bytes), packed structs get misaligned
+	 * pointers after the atomic field, causing wrong field offsets. */
 	#if MAX_RUNTIME_THREADS > 1
-	volatile unsigned int value;
+	volatile unsigned long value;
 	#else
-	unsigned int value;
+	unsigned long value;
 	#endif
 } _PACK_STRUCT;
